Ability to enter/exit rooms (read desc.)
The versions of Slick2d and Tiled that I use are incompatible in that
the program will crash unless object layers are assigned width and
height properties. These values are removed by Tiled right now, so I
have to enter width and height properties manually into object layers
after modifying a map. I will try to find a fix to this.